split slirp4netns.c; no substantial code change
