+++ Makefile.am
@@ -14,6 +14,7 @@ EXTRA_DIST = \
 	tests/common.sh \
 	slirp4netns.h \
 	api.h \
+	sandbox.h \
 	vendor/libslirp/COPYRIGHT \
 	vendor/libslirp/README.md \
 	vendor/libslirp/src/bootp.h \
@@ -91,7 +92,7 @@ libslirp_a_CFLAGS = $(AM_CFLAGS) -I$(abs_top_builddir)/vendor/libslirp/src
 libparson_a_CFLAGS = $(AM_CFLAGS) -I$(abs_top_builddir)/vendor/parson
 libparson_a_SOURCES = vendor/parson/parson.c
 
-slirp4netns_SOURCES = main.c slirp4netns.c api.c
+slirp4netns_SOURCES = main.c slirp4netns.c api.c sandbox.c
 slirp4netns_LDADD = libslirp.a libparson.a @GLIB_LIBS@ -lpthread
 man1_MANS = slirp4netns.1
